DATAMONGO-1231 - Move Auditor to package private usage only
