+++ src/main/java/org/springframework/data/domain/AuditorAware.java
@@ -22,7 +22,6 @@ import java.util.Optional;
  *
  * @param <T> the type of the auditing instance.
  * @author Oliver Gierke
- * @author Christoph Strobl
  */
 public interface AuditorAware<T> {
 
@@ -33,13 +32,4 @@ public interface AuditorAware<T> {
 	 */
 	Optional<T> getCurrentAuditor();
 
-	/**
-	 * Returns the current {@link Auditor} of the application.
-	 *
-	 * @return the current {@link Auditor}, {@link Auditor#none() none} if not present.
-	 * @since 2.4
-	 */
-	default Auditor<T> getAuditor() {
-		return Auditor.ofOptional(getCurrentAuditor());
-	}
 }
