+++ src/test/java/org/springframework/data/auditing/AuditorUnitTests.java
@@ -20,7 +20,6 @@ import static org.assertj.core.api.Assertions.*;
 import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.data.domain.Auditor;
 
 /**
  * @author Christoph Strobl
